test: Move create_islock to DashTestFramework in test_framework.py
Just because it's used the same way in two files
