+++ test/functional/feature_llmq_is_cl_conflicts.py
@@ -285,25 +285,5 @@ class LLMQ_IS_CL_Conflicts(DashTestFramework):
         clsig = msg_clsig(height, block.sha256, hex_str_to_bytes(recSig['sig']))
         return clsig
 
-    def create_islock(self, hextx):
-        tx = FromHex(CTransaction(), hextx)
-        tx.rehash()
-
-        request_id_buf = ser_string(b"islock") + ser_compact_size(len(tx.vin))
-        inputs = []
-        for txin in tx.vin:
-            request_id_buf += txin.prevout.serialize()
-            inputs.append(txin.prevout)
-        request_id = hash256(request_id_buf)[::-1].hex()
-        message_hash = tx.hash
-
-        for mn in self.mninfo:
-            mn.node.quorum('sign', 100, request_id, message_hash)
-
-        recSig = self.get_recovered_sig(request_id, message_hash)
-        islock = msg_islock(inputs, tx.sha256, hex_str_to_bytes(recSig['sig']))
-        return islock
-
-
 if __name__ == '__main__':
     LLMQ_IS_CL_Conflicts().main()
